Create dummy columns from logicals or factors in a data frame

This function is deprecated because partition() is more general and can be used to create dummy columns as well.

Create dummy logical columns from selected columns of the data frame. Dummy columns may be created for logical or factor columns as follows:

for logical column col, a pair of columns is created named col=T and col=F where the former (resp. latter) is equal to the original (resp. negation of the original);

for factor column col, a new logical column is created for each level l of the factor col and named as col=l with a value set to TRUE wherever the original column is equal to l.

Usage

dichotomize(.data, what = everything(), ..., .keep = FALSE, .other = FALSE)

Arguments

.data

a data frame to be processed

what

a tidyselect expression (see tidyselect syntax) selecting the columns to be processed

...

further tidyselect expressions for selecting the columns to be processed

.keep

whether to keep the original columns. If FALSE, the original columns are removed from the result.

.other

whether to put into result the rest of columns that were not specified for dichotomization in what argument.

Value

A tibble with selected columns replaced with dummy columns.

Author

Michal Burda

Examples

dichotomize(CO2, Plant:Treatment, .other = TRUE)
#> Warning: `dichotomize()` was deprecated in nuggets 1.4.0.
#> ℹ Please use `partition()` instead.
#> # A tibble: 84 × 18
#>     conc uptake `Plant=Qn1` `Plant=Qn2` `Plant=Qn3` `Plant=Qc1` `Plant=Qc3`
#>    <dbl>  <dbl> <lgl>       <lgl>       <lgl>       <lgl>       <lgl>      
#>  1    95   16   TRUE        FALSE       FALSE       FALSE       FALSE      
#>  2   175   30.4 TRUE        FALSE       FALSE       FALSE       FALSE      
#>  3   250   34.8 TRUE        FALSE       FALSE       FALSE       FALSE      
#>  4   350   37.2 TRUE        FALSE       FALSE       FALSE       FALSE      
#>  5   500   35.3 TRUE        FALSE       FALSE       FALSE       FALSE      
#>  6   675   39.2 TRUE        FALSE       FALSE       FALSE       FALSE      
#>  7  1000   39.7 TRUE        FALSE       FALSE       FALSE       FALSE      
#>  8    95   13.6 FALSE       TRUE        FALSE       FALSE       FALSE      
#>  9   175   27.3 FALSE       TRUE        FALSE       FALSE       FALSE      
#> 10   250   37.1 FALSE       TRUE        FALSE       FALSE       FALSE      
#> # ℹ 74 more rows
#> # ℹ 11 more variables: `Plant=Qc2` <lgl>, `Plant=Mn3` <lgl>, `Plant=Mn2` <lgl>,
#> #   `Plant=Mn1` <lgl>, `Plant=Mc2` <lgl>, `Plant=Mc3` <lgl>, `Plant=Mc1` <lgl>,
#> #   `Type=Quebec` <lgl>, `Type=Mississippi` <lgl>,
#> #   `Treatment=nonchilled` <lgl>, `Treatment=chilled` <lgl>
# ->
partition(CO2, Plant:Treatment)
#> # A tibble: 84 × 18
#>     conc uptake `Plant=Qn1` `Plant=Qn2` `Plant=Qn3` `Plant=Qc1` `Plant=Qc3`
#>    <dbl>  <dbl> <lgl>       <lgl>       <lgl>       <lgl>       <lgl>      
#>  1    95   16   TRUE        FALSE       FALSE       FALSE       FALSE      
#>  2   175   30.4 TRUE        FALSE       FALSE       FALSE       FALSE      
#>  3   250   34.8 TRUE        FALSE       FALSE       FALSE       FALSE      
#>  4   350   37.2 TRUE        FALSE       FALSE       FALSE       FALSE      
#>  5   500   35.3 TRUE        FALSE       FALSE       FALSE       FALSE      
#>  6   675   39.2 TRUE        FALSE       FALSE       FALSE       FALSE      
#>  7  1000   39.7 TRUE        FALSE       FALSE       FALSE       FALSE      
#>  8    95   13.6 FALSE       TRUE        FALSE       FALSE       FALSE      
#>  9   175   27.3 FALSE       TRUE        FALSE       FALSE       FALSE      
#> 10   250   37.1 FALSE       TRUE        FALSE       FALSE       FALSE      
#> # ℹ 74 more rows
#> # ℹ 11 more variables: `Plant=Qc2` <lgl>, `Plant=Mn3` <lgl>, `Plant=Mn2` <lgl>,
#> #   `Plant=Mn1` <lgl>, `Plant=Mc2` <lgl>, `Plant=Mc3` <lgl>, `Plant=Mc1` <lgl>,
#> #   `Type=Quebec` <lgl>, `Type=Mississippi` <lgl>,
#> #   `Treatment=nonchilled` <lgl>, `Treatment=chilled` <lgl>