Search for contrast patterns

Contrast patterns are a generalization of association rules that allow for the specification of a condition under which there is a significant difference between two numeric variables. The contrast is computed using a statistical test, which is specified by the method argument. The function computes the contrast between all pairs of variables, where the first variable is specified by the xvars argument and the second variable is specified by the yvars argument. The contrast is computed in sub-data corresponding to conditions generated from the condition columns.

Usage

dig_contrasts(
  x,
  condition = where(is.logical),
  xvars = where(is.numeric),
  yvars = where(is.numeric),
  method = "t",
  alternative = "two.sided",
  min_length = 0L,
  max_length = Inf,
  min_support = 0,
  max_p_value = 0.05,
  threads = 1,
  ...
)

Arguments

x

a matrix or data frame with data to search in.

condition

a tidyselect expression (see tidyselect syntax) specifying the columns to use as condition predicates

xvars

a tidyselect expression (see tidyselect syntax) specifying the columns to use for computation of contrasts

yvars

a tidyselect expression (see tidyselect syntax) specifying the columns to use for computation of contrasts

method

a character string indicating which contrast to compute. One of "t", "wilcox", or "var". "t" (resp. "wilcos") compute a parametric (resp. non-parametric) test on equality in position, and "var" performs the F-test on equality of variance.

alternative

indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less". "greater" corresponds to positive association, "less" to negative association.

min_length

the minimum size (the minimum number of predicates) of the condition to be generated (must be greater or equal to 0). If 0, the empty condition is generated in the first place.

max_length

The maximum size (the maximum number of predicates) of the condition to be generated. If equal to Inf, the maximum length of conditions is limited only by the number of available predicates.

min_support

the minimum support of a condition to trigger the callback function for it. The support of the condition is the relative frequency of the condition in the dataset x. For logical data, it equals to the relative frequency of rows such that all condition predicates are TRUE on it. For numerical (double) input, the support is computed as the mean (over all rows) of multiplications of predicate values.

max_p_value

the maximum p-value of a test for the pattern to be considered significant. If the p-value of the test is greater than max_p_value, the pattern is not included in the result.

threads

the number of threads to use for parallel computation.

...

Further arguments passed to the underlying test function (t.test(), wilcox.test(), or var.test() accordingly to the selected method).

Value

A tibble with found rules.

See also

Author

Michal Burda

Examples

crispCO2 <- partition(CO2, Plant:Treatment)
dig_contrasts(crispCO2,
              condition = where(is.logical),
              xvars = conc,
              yvars = uptake,
              method = "t",
              min_support = 0.1)
#> # A tibble: 9 × 15
#>   condition support xvar  yvar  estimate_x estimate_y t_statistic    df  p_value
#>   <chr>       <dbl> <chr> <chr>      <dbl>      <dbl>       <dbl> <dbl>    <dbl>
#> 1 {}           1    conc  upta…        435       27.2       12.6   83.2 5.02e-21
#> 2 {Type=Qu…    0.5  conc  upta…        435       33.5        8.73  41.1 6.54e-11
#> 3 {Type=Mi…    0.5  conc  upta…        435       20.9        9.01  41.1 2.81e-11
#> 4 {Treatme…    0.5  conc  upta…        435       30.6        8.80  41.1 5.38e-11
#> 5 {Treatme…    0.5  conc  upta…        435       23.8        8.95  41.1 3.40e-11
#> 6 {Treatme…    0.25 conc  upta…        435       31.8        6.13  20.0 5.43e- 6
#> 7 {Treatme…    0.25 conc  upta…        435       15.8        6.37  20.0 3.22e- 6
#> 8 {Treatme…    0.25 conc  upta…        435       35.3        6.07  20.0 6.12e- 6
#> 9 {Treatme…    0.25 conc  upta…        435       26.0        6.22  20.0 4.49e- 6
#> # ℹ 6 more variables: rows <int>, conf_int_lo <dbl>, conf_int_hi <dbl>,
#> #   stderr <dbl>, alternative <chr>, method <chr>