Format a vector of predicates into a string with a condition

Function takes a character vector of predicates and returns a formatted condition. The format of the condition is a string with predicates separated by commas and enclosed in curly braces.

Usage

format_condition(condition)

Arguments

condition

a character vector of predicates to be formatted

Value

a character scalar with a formatted condition

Author

Michal Burda

Examples

format_condition(NULL)              # returns {}
#> [1] "{}"
format_condition(c("a", "b", "c"))  # returns {a,b,c}
#> [1] "{a,b,c}"