nuggets: Extensible Framework for Data Pattern Exploration

A framework for systematic exploration of association rules (Agrawal et al., 1994, https://www.vldb.org/conf/1994/P487.PDF), contrast patterns (Chen, 2022, doi:10.48550/arXiv.2209.13556 ), emerging patterns (Dong et al., 1999, doi:10.1145/312129.312191 ), subgroup discovery (Atzmueller, 2015, doi:10.1002/widm.1144 ), and conditional correlations (Hájek, 1978, doi:10.1007/978-3-642-66943-9 ). User-defined functions may also be supplied to guide custom pattern searches. Supports both crisp (Boolean) and fuzzy data. Generates candidate conditions expressed as elementary conjunctions, evaluates them on a dataset, and inspects the induced sub-data for statistical, logical, or structural properties such as associations, correlations, or contrasts. Includes methods for visualization of logical structures and supports interactive exploration through integrated Shiny applications.

See also

Author

Maintainer: Michal Burda michal.burda@osu.cz (ORCID)